+++ src/components/Sidebar/NavItem/Navitem.css.ts
@@ -1,0 +1,7 @@
+import {
+  label,
+  navItem,
+  navItemIcon,
+} from '@/components/Sidebar/sidebarCommon.css.ts';
+
+export { navItem, navItemIcon, label };
