+++ src/components/Sidebar/NavItem/index.ts
@@ -1,0 +1,1 @@
+export * from './NavItem';
